requirements.txt
# Super Swing Trading Platform - Complete Dependencies
# Python >= 3.9 required

# ============================================
# Core Data Processing
# ============================================
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0
pytz>=2023.3
scipy>=1.11.0

# ============================================
# Technical Analysis
# ============================================
# TA-Lib>=0.4.28
# pandas-ta>=0.3.14b0
yfinance>=0.2.28
nsepy>=0.8

# ============================================
# Machine Learning
# ============================================
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# ============================================
# Database
# ============================================
sqlalchemy>=2.0.0

# ============================================
# Visualization
# ============================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ============================================
# Web Framework & API
# ============================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
pydantic>=2.0.0
email-validator>=2.1.0
flask>=3.0.0
flask-cors>=4.0.0

# ============================================
# Trading & Broker APIs
# ============================================
kiteconnect>=4.2.0

# ============================================
# AI & LLM Integration
# ============================================
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.36.0

# ============================================
# Price Prediction
# ============================================
statsforecast>=1.6.0
fugue>=0.8.0

# ============================================
# Telegram Integration
# ============================================
telethon>=1.28.0
nest_asyncio>=1.5.0

# ============================================
# CLI & User Interface
# ============================================
click>=8.1.0
tqdm>=4.65.0
colorlog>=6.7.0

# ============================================
# Authentication & Security
# ============================================
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0

# ============================================
# HTTP & Networking
# ============================================
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# ============================================
# File Processing
# ============================================
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# ============================================
# Configuration & Environment
# ============================================
python-dotenv>=1.0.0
pyyaml>=6.0.1


# ============================================
# Testing
# ============================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# ============================================
# Utilities
# ============================================
joblib>=1.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ============================================
# Logging & Monitoring
# ============================================
loguru>=0.7.0

# ============================================
# Optional: For development
# ============================================
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
# ipython>=8.12.0
